Remove all classes and start over with a BaseCollection
This collection should have all of the 'basic' methods allowed to be
performed on a collection of objects. For instance: map.

We can work on providing a base 'ValueType' for each primitive value
that should be collectable by this class. For instance, a string *is*
technically iterable since we could iterate over the contents of the
string by using a for loop.

Numbers and booleans, however, are not iterable so we should not allow
them to be used as values of this class.

Finally, we don't have to port **every** lodash function. Just sensible
ones. For instance, the boolean and mathematical logic is probably a
negligible feature of Lodash nowadays and PHP is stronger in the field
of 'types' than it used to be. This makes those functions less desirable
since they would only serve to bloat the library.

Refocusing and rebranding Phlash to be a quick and easy to use
collection library should be the primary focus.

Also of note: Phlash should not devolve to simply being a competitor to
Laravel's collections. Phlash should evolve in its own direction with
thoughtful features so as to be an alternative approach to Laravel's collections.

That being said, the new approach of Phlash is to provide several
high-value functions with a reasonable and easy to remember API.